Force int type (Closes: #7341).
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@25605 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ webfilter-base/hier/usr/lib/python2.5/reports/node/untangle_base_webfilter.py
@@ -1310,7 +1310,7 @@ class WebFilterDetailDomains(DetailSection):
 
         sql = """\
 SELECT regexp_replace(host, E'.*?([^.]+\.[^.]+)(:[0-9]+)?$', E'\\\\1') AS domain,
-       count(*) AS count, COALESCE(sum(s2c_content_length) / 10^6, 0)
+       count(*) AS count, COALESCE(sum(s2c_content_length) / 10^6, 0)::int
 FROM reports.n_http_events
 WHERE regexp_replace(host, E'[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+(:[0-9]+)?', '') != ''
 AND time_stamp >= %s AND time_stamp < %s
